+++ src/serializer.rs
@@ -7,54 +7,86 @@ use crate::fullname::FullnameSerializer;
 use crate::name::NameId;
 use crate::namespace::NamespaceId;
 use crate::prefix::PrefixId;
-use crate::xmlvalue::{Comment, Element, ProcessingInstruction, Text, ToNamespace, Value};
+use crate::xmlvalue::{
+    Comment, Element, ProcessingInstruction, Text, ToNamespace, Value, ValueType,
+};
 use crate::xotdata::{Node, Xot};
 
 pub(crate) struct Serializer<'a, W: SerializerWriter> {
+    xot: &'a Xot<'a>,
     writer: &'a mut W,
-    extra_prefixes: ToNamespace,
+    fullname_serializer: FullnameSerializer<'a>,
 }
 
 impl<'a, W: SerializerWriter> Serializer<'a, W> {
-    pub(crate) fn new(writer: &'a mut W, extra_prefixes: ToNamespace) -> Self {
+    /// Create a new serializer from a serializer writer.
+
+    pub fn new(xot: &'a Xot<'a>, writer: &'a mut W) -> Self {
+        let fullname_serializer = FullnameSerializer::new(xot);
         Self {
+            xot,
             writer,
-            extra_prefixes,
+            fullname_serializer,
         }
     }
 
-    pub(crate) fn serialize_node(&mut self, xot: &Xot, node: Node) -> Result<(), Error> {
-        for edge in xot.traverse(node) {
+    fn get_extra_prefixes(&self, node: Node) -> ToNamespace {
+        // collect namespace prefixes for all ancestors of the fragment
+        if let Some(parent) = self.xot.parent(node) {
+            if self.xot.value_type(parent) != ValueType::Root {
+                self.xot.to_namespace_in_scope(parent)
+            } else {
+                ToNamespace::new()
+            }
+        } else {
+            ToNamespace::new()
+        }
+    }
+
+    /// Serialize a node and all its children using the writer.
+    ///
+    /// The writer controls what happens with the serialized data.
+    pub(crate) fn serialize_node(&mut self, node: Node) -> Result<(), Error> {
+        for edge in self.xot.traverse(node) {
             match edge {
                 NodeEdge::Start(current_node) => {
-                    self.handle_edge_start(xot, node, current_node)?;
+                    self.handle_edge_start(node, current_node)?;
                 }
                 NodeEdge::End(current_node) => {
-                    self.handle_edge_end(xot, current_node)?;
+                    self.handle_edge_end(current_node)?;
                 }
             }
         }
         Ok(())
     }
 
-    fn handle_edge_start(&mut self, xot: &Xot, top_node: Node, node: Node) -> Result<(), Error> {
-        let value = xot.value(node);
+    fn handle_edge_start(&mut self, top_node: Node, node: Node) -> Result<(), Error> {
+        let value = self.xot.value(node);
         match value {
             Value::Root => {}
             Value::Element(element) => {
-                self.writer.write_start_tag_open(node, element)?;
+                if node == top_node {
+                    let extra_prefixes = self.get_extra_prefixes(node);
+                    self.fullname_serializer.push(&extra_prefixes);
+                    // since this will always be at the bottom, it doesn't need to be popped
+                }
+                self.fullname_serializer
+                    .push(&element.namespace_info.to_namespace);
+                let fullname = self.fullname_serializer.fullname_or_err(element.name_id)?;
+                self.writer
+                    .write_start_tag_open(node, element, fullname.as_ref())?;
 
                 // serialize any extra prefixes if this is the top element of
                 // a fragment and they aren't declared already
                 if node == top_node {
-                    for (prefix_id, namespace_id) in &self.extra_prefixes {
-                        if !element.namespace_info.to_namespace.contains_key(prefix_id) {
+                    for (prefix_id, namespace_id) in self.get_extra_prefixes(node) {
+                        if !element.namespace_info.to_namespace.contains_key(&prefix_id) {
                             self.writer.write_space()?;
                             self.writer.write_namespace_declaration(
                                 node,
                                 element,
-                                *prefix_id,
-                                *namespace_id,
+                                prefix_id,
+                                namespace_id,
                             )?;
                         }
                     }
@@ -72,8 +104,14 @@ impl<'a, W: SerializerWriter> Serializer<'a, W> {
 
                 for (name_id, value) in element.attributes() {
                     self.writer.write_space()?;
-                    self.writer
-                        .write_attribute(node, element, *name_id, value)?;
+                    let fullname = self.fullname_serializer.fullname_attr_or_err(*name_id)?;
+                    self.writer.write_attribute(
+                        node,
+                        element,
+                        *name_id,
+                        value,
+                        fullname.as_ref(),
+                    )?;
                 }
 
                 self.writer.write_start_tag_close(node, element)?;
@@ -91,19 +129,50 @@ impl<'a, W: SerializerWriter> Serializer<'a, W> {
         Ok(())
     }
 
-    fn handle_edge_end(&mut self, xot: &Xot, node: Node) -> Result<(), Error> {
-        let value = xot.value(node);
+    fn handle_edge_end(&mut self, node: Node) -> Result<(), Error> {
+        let value = self.xot.value(node);
         if let Value::Element(element) = value {
-            self.writer.write_end_tag(node, element)?;
+            let fullname = self.fullname_serializer.fullname_or_err(element.name_id)?;
+            self.writer
+                .write_end_tag(node, element, fullname.as_ref())?;
+            self.fullname_serializer
+                .pop(&element.namespace_info.to_namespace);
         }
         Ok(())
     }
 }
 
+/// A trait for writing XML.
+///
+/// The serializer invokes methods on this trait to actually write the XML.
+///
+/// This is an advanced trait that can be used with
+/// [`Xot::serialize_with_writer`], [`Xot::serialize_node_with_writer`] and
+/// [`Xot::serialize_or_missing_prefix_with_writer`].
+///
+/// Usually you'd implement this trait in your own struct by using
+/// [`xot::StringWriter`](`crate::serializer::StringWriter`), which does the
+/// standard XML serialization. The default serialization process can be
+/// understood as using
+/// [`xot::StringWriter`](`crate::serializer::StringWriter`) to implement each
+/// of these methods and pass the resulting strings off to an output buffer.
+/// But in your implementation of this trait you can decide to output
+/// additional content, or to filter out particular strings and not write them
+/// at all.
 pub trait SerializerWriter {
-    fn write_start_tag_open(&mut self, node: Node, element: &Element) -> Result<(), Error>;
+    /// Write the start tag opening: e.g `<foo`.
+    fn write_start_tag_open(
+        &mut self,
+        node: Node,
+        element: &Element,
+        fullname: &str,
+    ) -> Result<(), Error>;
+    /// Write the start tag closing, e.g. `>`, or `/>` if the element is empty.
     fn write_start_tag_close(&mut self, node: Node, element: &Element) -> Result<(), Error>;
-    fn write_end_tag(&mut self, node: Node, element: &Element) -> Result<(), Error>;
+    /// Write the end tag, e.g. `</foo>`, or nothing if the element is empty.
+    fn write_end_tag(&mut self, node: Node, element: &Element, fullname: &str)
+        -> Result<(), Error>;
+    /// Write a namespace declaration, e.g. `xmlns:foo="http://example.com"`.
     fn write_namespace_declaration(
         &mut self,
         node: Node,
@@ -111,20 +180,26 @@ pub trait SerializerWriter {
         prefix_id: PrefixId,
         namespace_id: NamespaceId,
     ) -> Result<(), Error>;
+    /// Write an attribute, e.g. `foo="bar"`.
     fn write_attribute(
         &mut self,
         node: Node,
         element: &Element,
         name_id: NameId,
         value: &str,
+        fullname: &str,
     ) -> Result<(), Error>;
+    /// Write text, e.g `foo`.
     fn write_text(&mut self, node: Node, text: &Text) -> Result<(), Error>;
+    /// Write a comment, e.g. `<!-- foo -->`.
     fn write_comment(&mut self, node: Node, comment: &Comment) -> Result<(), Error>;
+    /// Write a processing instruction, e.g. `<?foo bar?>`.
     fn write_processing_instruction(
         &mut self,
         node: Node,
         pi: &ProcessingInstruction,
     ) -> Result<(), Error>;
+    /// Write a space, e.g. ` `.
     fn write_space(&mut self) -> Result<(), Error>;
 }
 
@@ -160,18 +235,22 @@ impl Write for LastPushedWrite {
     }
 }
 
-struct StringWriter<'a> {
+/// Return strings for written XML.
+///
+/// This is like a SerializerWriter, but returns strings. This is useful if you
+/// want to create your own custom
+/// [`xot::SerializerWriter`](`crate::serializer::SerializerWriter`) trait
+/// implementation and use this inside it. You can decide whether to write a
+/// string at all, and also output additional content.
+pub struct StringWriter<'a> {
     serializer_writer: XmlSerializerWriter<'a, LastPushedWrite>,
 }
 
 impl<'a> StringWriter<'a> {
-    fn new(xot: &'a Xot, extra_prefixes: &ToNamespace) -> StringWriter<'a> {
+    /// Create a new StringWriter.
+    pub fn new(xot: &'a Xot) -> StringWriter<'a> {
         StringWriter {
-            serializer_writer: XmlSerializerWriter::new(
-                xot,
-                LastPushedWrite::new(),
-                extra_prefixes,
-            ),
+            serializer_writer: XmlSerializerWriter::new(xot, LastPushedWrite::new()),
         }
     }
 
@@ -183,23 +262,39 @@ impl<'a> StringWriter<'a> {
         self.serializer_writer.w.get()
     }
 
-    fn get_start_tag_open(&mut self, node: Node, element: &Element) -> Result<String, Error> {
+    /// Get opening of start tag
+    pub fn get_start_tag_open(
+        &mut self,
+        node: Node,
+        element: &Element,
+        fullname: &str,
+    ) -> Result<String, Error> {
         self.clear();
-        self.serializer_writer.write_start_tag_open(node, element)?;
+        self.serializer_writer
+            .write_start_tag_open(node, element, fullname)?;
         Ok(self.get())
     }
-    fn get_start_tag_close(&mut self, node: Node, element: &Element) -> Result<String, Error> {
+    /// Get end of start tag
+    pub fn get_start_tag_close(&mut self, node: Node, element: &Element) -> Result<String, Error> {
         self.clear();
         self.serializer_writer
             .write_start_tag_close(node, element)?;
         Ok(self.get())
     }
-    fn get_end_tag(&mut self, node: Node, element: &Element) -> Result<String, Error> {
+    /// Get end tag
+    pub fn get_end_tag(
+        &mut self,
+        node: Node,
+        element: &Element,
+        fullname: &str,
+    ) -> Result<String, Error> {
         self.clear();
-        self.serializer_writer.write_end_tag(node, element)?;
+        self.serializer_writer
+            .write_end_tag(node, element, fullname)?;
         Ok(self.get())
     }
-    fn get_namespace_declaration(
+    /// Get namespace declaration
+    pub fn get_namespace_declaration(
         &mut self,
         node: Node,
         element: &Element,
@@ -215,29 +310,34 @@ impl<'a> StringWriter<'a> {
         )?;
         Ok(self.get())
     }
-    fn get_attribute(
+    /// Get attribute
+    pub fn get_attribute(
         &mut self,
         node: Node,
         element: &Element,
         name_id: NameId,
         value: &str,
+        fullname: &str,
     ) -> Result<String, Error> {
         self.clear();
         self.serializer_writer
-            .write_attribute(node, element, name_id, value)?;
+            .write_attribute(node, element, name_id, value, fullname)?;
         Ok(self.get())
     }
-    fn get_text(&mut self, node: Node, text: &Text) -> Result<String, Error> {
+    /// Get text
+    pub fn get_text(&mut self, node: Node, text: &Text) -> Result<String, Error> {
         self.clear();
         self.serializer_writer.write_text(node, text)?;
         Ok(self.get())
     }
-    fn get_comment(&mut self, node: Node, comment: &Comment) -> Result<String, Error> {
+    /// Get comment
+    pub fn get_comment(&mut self, node: Node, comment: &Comment) -> Result<String, Error> {
         self.clear();
         self.serializer_writer.write_comment(node, comment)?;
         Ok(self.get())
     }
-    fn get_processing_instruction(
+    /// Get processing instruction
+    pub fn get_processing_instruction(
         &mut self,
         node: Node,
         pi: &ProcessingInstruction,
@@ -247,46 +347,35 @@ impl<'a> StringWriter<'a> {
             .write_processing_instruction(node, pi)?;
         Ok(self.get())
     }
-    fn get_space(&mut self) -> Result<String, Error> {
+    /// Get space
+    pub fn get_space(&mut self) -> Result<String, Error> {
         Ok(" ".to_string())
     }
 }
 
 pub(crate) struct XmlSerializerWriter<'a, W: Write> {
     xot: &'a Xot<'a>,
-    fullname_serializer: FullnameSerializer<'a>,
     w: W,
 }
 
 impl<'a, W: Write> XmlSerializerWriter<'a, W> {
-    pub(crate) fn new(
-        xot: &'a Xot<'a>,
-        w: W,
-        extra_prefixes: &ToNamespace,
-    ) -> XmlSerializerWriter<'a, W> {
-        let fullname_serializer =
-            FullnameSerializer::with_to_namespace(extra_prefixes.clone(), xot);
-        XmlSerializerWriter {
-            xot,
-            fullname_serializer,
-            w,
-        }
+    pub(crate) fn new(xot: &'a Xot<'a>, w: W) -> XmlSerializerWriter<'a, W> {
+        XmlSerializerWriter { xot, w }
     }
 }
 
 impl<'a, W: Write> SerializerWriter for XmlSerializerWriter<'a, W> {
-    fn write_start_tag_open(&mut self, node: Node, element: &Element) -> Result<(), Error> {
-        self.fullname_serializer
-            .push(&element.namespace_info.to_namespace);
-        write!(
-            self.w,
-            "<{}",
-            self.fullname_serializer.fullname_or_err(element.name())?
-        )?;
+    fn write_start_tag_open(
+        &mut self,
+        _node: Node,
+        _element: &Element,
+        fullname: &str,
+    ) -> Result<(), Error> {
+        write!(self.w, "<{}", fullname)?;
         Ok(())
     }
 
-    fn write_start_tag_close(&mut self, node: Node, element: &Element) -> Result<(), Error> {
+    fn write_start_tag_close(&mut self, node: Node, _element: &Element) -> Result<(), Error> {
         if self.xot.first_child(node).is_none() {
             write!(self.w, "/>")?;
         } else {
@@ -295,20 +384,22 @@ impl<'a, W: Write> SerializerWriter for XmlSerializerWriter<'a, W> {
         Ok(())
     }
 
-    fn write_end_tag(&mut self, node: Node, element: &Element) -> Result<(), Error> {
+    fn write_end_tag(
+        &mut self,
+        node: Node,
+        _element: &Element,
+        fullname: &str,
+    ) -> Result<(), Error> {
         if self.xot.first_child(node).is_some() {
-            let fullname = self.fullname_serializer.fullname_or_err(element.name())?;
             write!(self.w, "</{}>", fullname)?;
         }
-        self.fullname_serializer
-            .pop(&element.namespace_info.to_namespace);
         Ok(())
     }
 
     fn write_namespace_declaration(
         &mut self,
-        node: Node,
-        element: &Element,
+        _node: Node,
+        _element: &Element,
         prefix_id: PrefixId,
         namespace_id: NamespaceId,
     ) -> Result<(), Error> {
@@ -328,12 +419,12 @@ impl<'a, W: Write> SerializerWriter for XmlSerializerWriter<'a, W> {
 
     fn write_attribute(
         &mut self,
-        node: Node,
-        element: &Element,
-        name_id: NameId,
+        _node: Node,
+        _element: &Element,
+        _name_id: NameId,
         value: &str,
+        fullname: &str,
     ) -> Result<(), Error> {
-        let fullname = self.fullname_serializer.fullname_attr_or_err(name_id)?;
         write!(
             self.w,
             "{}=\"{}\"",
@@ -343,19 +434,19 @@ impl<'a, W: Write> SerializerWriter for XmlSerializerWriter<'a, W> {
         Ok(())
     }
 
-    fn write_text(&mut self, node: Node, text: &Text) -> Result<(), Error> {
+    fn write_text(&mut self, _node: Node, text: &Text) -> Result<(), Error> {
         write!(self.w, "{}", serialize_text(text.get().into()))?;
         Ok(())
     }
 
-    fn write_comment(&mut self, node: Node, comment: &Comment) -> Result<(), Error> {
+    fn write_comment(&mut self, _node: Node, comment: &Comment) -> Result<(), Error> {
         write!(self.w, "<!--{}-->", comment.get())?;
         Ok(())
     }
 
     fn write_processing_instruction(
         &mut self,
-        node: Node,
+        _node: Node,
         processing_instruction: &ProcessingInstruction,
     ) -> Result<(), Error> {
         if let Some(data) = processing_instruction.data() {
@@ -409,10 +500,9 @@ mod tests {
         }
 
         let mut w = RecordingWrite::new();
-        let extra_prefixes = ToNamespace::new();
-        let mut writer = XmlSerializerWriter::new(&xot, &mut w, &extra_prefixes);
-        let mut serializer = Serializer::new(&mut writer, extra_prefixes);
-        serializer.serialize_node(&xot, doc).unwrap();
+        let mut writer = XmlSerializerWriter::new(&xot, &mut w);
+        let mut serializer = Serializer::new(&xot, &mut writer);
+        serializer.serialize_node(doc).unwrap();
         assert_eq!(w.data.join(""), r#"<doc><a/></doc>"#);
     }
 
@@ -435,8 +525,8 @@ mod tests {
         }
 
         impl<'a> StyleWriter<'a> {
-            fn new(xot: &'a Xot<'a>, extra_prefixes: &ToNamespace) -> StyleWriter<'a> {
-                let inner_writer = StringWriter::new(xot, extra_prefixes);
+            fn new(xot: &'a Xot<'a>) -> StyleWriter<'a> {
+                let inner_writer = StringWriter::new(xot);
                 StyleWriter {
                     data: Vec::new(),
                     inner_writer,
@@ -446,8 +536,15 @@ mod tests {
         }
 
         impl<'a> SerializerWriter for StyleWriter<'a> {
-            fn write_start_tag_open(&mut self, node: Node, element: &Element) -> Result<(), Error> {
-                let text = self.inner_writer.get_start_tag_open(node, element)?;
+            fn write_start_tag_open(
+                &mut self,
+                node: Node,
+                element: &Element,
+                fullname: &str,
+            ) -> Result<(), Error> {
+                let text = self
+                    .inner_writer
+                    .get_start_tag_open(node, element, fullname)?;
                 let name_a = self.xot.name("a").unwrap();
                 if element.name() == name_a {
                     self.data.push(StyledText::StyleStart);
@@ -468,8 +565,13 @@ mod tests {
                 Ok(())
             }
 
-            fn write_end_tag(&mut self, node: Node, element: &Element) -> Result<(), Error> {
-                let text = self.inner_writer.get_end_tag(node, element)?;
+            fn write_end_tag(
+                &mut self,
+                node: Node,
+                element: &Element,
+                fullname: &str,
+            ) -> Result<(), Error> {
+                let text = self.inner_writer.get_end_tag(node, element, fullname)?;
                 let name_a = self.xot.name("a").unwrap();
                 if element.name() == name_a {
                     self.data.push(StyledText::Text(text));
@@ -503,10 +605,11 @@ mod tests {
                 element: &Element,
                 name_id: NameId,
                 value: &str,
+                fullname: &str,
             ) -> Result<(), Error> {
                 let text = self
                     .inner_writer
-                    .get_attribute(node, element, name_id, value)?;
+                    .get_attribute(node, element, name_id, value, fullname)?;
                 self.data.push(StyledText::Text(text));
                 Ok(())
             }
@@ -540,11 +643,9 @@ mod tests {
             }
         }
 
-        let extra_prefixes = ToNamespace::new();
-
-        let mut writer = StyleWriter::new(&xot, &extra_prefixes);
-        let mut serializer = Serializer::new(&mut writer, extra_prefixes);
-        serializer.serialize_node(&xot, doc).unwrap();
+        let mut writer = StyleWriter::new(&xot);
+        let mut serializer = Serializer::new(&xot, &mut writer);
+        serializer.serialize_node(doc).unwrap();
 
         let data = &writer.data;